// src/models/enums/logaccess.enum.ts
enum LOGTYPE {
  LOGIN = 1,
  LOGOUT = 2,
  GETGRADES = 3,
  GETGRADESPROGRESS = 4,
  GETLEVELS = 5,
  GETLEVELSPROGRESS = 6,
  GETLESSONS = 7,
  GETLESSONSPROGRESS = 8,
  POSTLEARNING = 9,
  GETLEARNINGPROGRESS = 10,
  GETALLLEARNINGS = 11,
  EXPORTLOG = 12,
  IMPORTSTUDENT = 13,
  IMPORTTEACHER = 14,
  IMPORTCONTENTS = 15,
  SUBMITPRACTICE = 16,
  SUBMITQUIZ = 17,
  SUBMITLEVELQUIZ = 18,
  EXPORTLOGFILES = 19,
  GETBASELINE = 20,
  GETLEVELQUIZ = 21,
  GETLESSONQUIZ = 22,
  GETLESSONPRACTICE = 23,
  GETCURRICULUM = 24,
}

export { LOGTYPE };


export const LOGDIR = 'logs';