// src/models/Sync.ts
export class Sync {

    curriculums: Array<any> = [];
    curriculumbaselines: Array<any> = [];
    baselinequestion: Array<any> = [];
    grades: Array<any> = [];
    levels: Array<any> = [];
    lessons: Array<any> = [];
    lessonlearnings: Array<any> = [];
    lessonpractices: Array<any> = [];
    lessonquizzes: Array<any> = [];
    lessonpracticequestions: Array<any> = [];
    lessonquizquestions: Array<any> = [];
    levelquizquestions: Array<any> = [];
    questions: Array<any> = [];
    documents: Array<any> = [];
    standards: Array<any> = [];
    schools: Array<any> = [];
    countries: Array<any> = [];
    lessonplans: Array<any> = [];
    subjects: Array<any> = [];
}